+++ ExtasyCompleteNavigation/Utilities/Extensions.swift
@@ -1,7 +1,0 @@
-//
-//  Extensions.swift
-//  ExtasyCompleteNavigation
-//
-//  Created by Vasil Borisov on 18.11.24.
-//
-